+++ src/id-verification/panels/SubmittedPanel.jsx
@@ -5,7 +5,7 @@ import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 
 import BasePanel from './BasePanel';
 
-import { IdVerificationContext } from '../IdVerificationContext';
+import IdVerificationContext from '../IdVerificationContext';
 import messages from '../IdVerification.messages';
 
 function SubmittedPanel(props) {
