+++ src/id-verification/panels/TakeIdPhotoPanel.jsx
@@ -5,7 +5,7 @@ import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 import { useNextPanelSlug } from '../routing-utilities';
 import BasePanel from './BasePanel';
 import Camera from '../Camera';
-import { IdVerificationContext } from '../IdVerificationContext';
+import IdVerificationContext from '../IdVerificationContext';
 
 import messages from '../IdVerification.messages';
 import CameraHelp from '../CameraHelp';
